add next-navigation
